requirements.txt
# Core dependencies
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.15.0

# Additional utilities
python-dotenv>=1.0.0
urllib3>=2.0.0

# Development dependencies (optional)
pytest>=7.4.0
unittest-xml-reporting>=3.2.0

# For data processing
numpy>=1.24.0

# For time handling
pytz>=2023.3